requirements.txt
numpy==1.24.4
pandas==2.0.3
tensorflow==2.9.2
matplotlib==3.7.5
seaborn==0.13.2
nltk==3.9.1
regex==2024.9.11
scikit-learn==1.3.2
from-root==1.3.0
google-cloud-storage==2.18.2
fastapi==0.78.0
uvicorn==0.18.3
jinja2==3.1.2
python-dotenv==1.0.1

-e .